+++ Runtime/Scripts/ImmersalSDK.cs
@@ -29,7 +29,7 @@ namespace Immersal
 	{
 		// SDK properties
 		
-		public static string sdkVersion = "2.0.2";
+		public static string sdkVersion = "2.0.3";
 		private static readonly string[] ServerList = new[] {"https://api.immersal.com", "https://immersal.hexagon.com.cn"};
 		public enum APIServer { DefaultServer = 0, ChinaServer = 1, CustomServer = 2 };
 		
@@ -153,7 +153,8 @@ namespace Immersal
         [Header("Events")]
         public UnityEvent OnInitializationComplete;
         public UnityEvent OnReset;
-
+        
+        [SerializeField]
         private string m_CustomServerUrl = "";
         private int m_LicenseLevel = -1;